Fix circular imports by splitting Connector out of __init__ file.
